Add compiler option detection in source code
This commit adds compiler option detection in the source code. To pass
an option to the compiler, use the command:
    %rlatex: cmd option
At this time, just compiler and output are the only two compiler
options that can be set in this manner.
unit-circle.tex is also added as an example on how to use this new
feature.